Big changes everywhere (see commit message)
* The users are not automatically displayed anymore in admin and sysadmin page page, but a search input has been added
* Add `editusers.js` for archive and destroy of users
* Drop the `mcrypt` extension requirement with the update of onelogin/php-saml to version 3.0.0
* Remove all views classes, remove `src/views`
* Add custom exceptions in `Elabftw\Exceptions` namespace. See "Exceptions handling" part of "Contributing" page on doc.elabftw.net
* Move Interfaces in own `src/` folder with own namespace
* Add controllers in `src/` folder to deal with requests and give response
* Throw custom exceptions instead of relying on bool return value to be checked for most models methods
* Typehint SessionInterface instead of Session
* Change Formkey class in Csrf and add Csrf checks everywhere
* Add Csrf class in App so it's always available
* Add two custom twig filters: getExt (|ext) and filesize (|filesize)
* Get rid of `userid` field on Users, only use userData
* Always check if the sql request was successful or throw exception instead of returning `req->execute()`
* Simplify and improve the code (get rid of some TODO and FIXME comments)
* Don't catch exceptions but let it bubble up
* Prefer custom exceptions
* Add Extensions class to group file extensions. It only contains arrays of file extensions
* Add MakeThumbnail class to create thumbnails for uploaded files (means less cluttered/better Uploads class)
* Fix issue with IDP not id of 1 (#1025) (add `active` attribute to IDPs
* Use better SQL for toggleLock (IF(locked = 1, 0, 1))
* Split controllers in normal and ajax
* Use Request instead of $_GET/$_POST
* Use RedirectResponse instead of header()
* 144 files changed, 5165 insertions(+), 3682 deletions(-)
…and it's not finished!

+++ src/models/Experiments.php
@@ -12,7 +12,9 @@ declare(strict_types=1);
 
 namespace Elabftw\Elabftw;
 
-use Exception;
+use Elabftw\Exceptions\DatabaseErrorException;
+use Elabftw\Exceptions\IllegalActionException;
+use Elabftw\Exceptions\ImproperActionException;
 use PDO;
 
 /**
@@ -55,7 +57,7 @@ class Experiments extends AbstractEntity
         $Templates = new Templates($this->Users);
 
         // do we want template ?
-        if ($tpl !== null) {
+        if ($tpl) {
             $Templates->setId($tpl);
             $templatesArr = $Templates->read();
             $title = $templatesArr['name'];
@@ -82,7 +84,7 @@ class Experiments extends AbstractEntity
             'status' => $this->getStatus(),
             'elabid' => $this->generateElabid(),
             'visibility' => $visibility,
-            'userid' => $this->Users->userid
+            'userid' => $this->Users->userData['userid']
         ));
         $newId = $this->Db->lastInsertId();
 
@@ -110,7 +112,9 @@ class Experiments extends AbstractEntity
             WHERE link_id = :link_id";
         $req = $this->Db->prepare($sql);
         $req->bindParam(':link_id', $itemId, PDO::PARAM_INT);
-        $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
         while ($data = $req->fetch()) {
             $this->setId((int) $data['item_id']);
             $this->canOrExplode('read');
@@ -124,16 +128,20 @@ class Experiments extends AbstractEntity
      * Update the status for an experiment
      *
      * @param int $status Id of the status
-     * @return bool
+     * @return void
      */
-    public function updateCategory(int $status): bool
+    public function updateCategory(int $status): void
     {
+        $this->canOrExplode('write');
+
         $sql = "UPDATE experiments SET status = :status WHERE id = :id AND locked = 0";
         $req = $this->Db->prepare($sql);
         $req->bindParam(':status', $status, PDO::PARAM_INT);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
 
-        return $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
     }
 
     /**
@@ -147,7 +155,9 @@ class Experiments extends AbstractEntity
         $sql = "SELECT is_timestampable FROM status WHERE id = :status;";
         $req = $this->Db->prepare($sql);
         $req->bindParam(':status', $currentStatus, PDO::PARAM_INT);
-        $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
         return (bool) $req->fetchColumn();
     }
 
@@ -156,9 +166,9 @@ class Experiments extends AbstractEntity
      *
      * @param string $responseTime the date of the timestamp
      * @param string $responsefilePath the file path to the timestamp token
-     * @return bool
+     * @return void
      */
-    public function updateTimestamp(string $responseTime, string $responsefilePath): bool
+    public function updateTimestamp(string $responseTime, string $responsefilePath): void
     {
         $sql = "UPDATE experiments SET
             locked = 1,
@@ -173,10 +183,12 @@ class Experiments extends AbstractEntity
         $req->bindParam(':when', $responseTime);
         // the date recorded in the db has to match the creation time of the timestamp token
         $req->bindParam(':longname', $responsefilePath);
-        $req->bindParam(':userid', $this->Users->userid, PDO::PARAM_INT);
+        $req->bindParam(':userid', $this->Users->userData['userid'], PDO::PARAM_INT);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
 
-        return $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
     }
 
     /**
@@ -193,7 +205,9 @@ class Experiments extends AbstractEntity
         $sql = 'SELECT id FROM status WHERE is_default = true AND team = :team LIMIT 1';
         $req = $this->Db->prepare($sql);
         $req->bindParam(':team', $this->Users->userData['team'], PDO::PARAM_INT);
-        $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
         $status = $req->fetchColumn();
 
         // if there is no is_default status
@@ -202,7 +216,9 @@ class Experiments extends AbstractEntity
             $sql = 'SELECT id FROM status WHERE team = :team LIMIT 1';
             $req = $this->Db->prepare($sql);
             $req->bindParam(':team', $this->Users->userData['team'], PDO::PARAM_INT);
-            $req->execute();
+            if ($req->execute() !== true) {
+                throw new DatabaseErrorException('Error while executing SQL query.');
+            }
             $status = $req->fetchColumn();
         }
         return (int) $status;
@@ -227,11 +243,6 @@ class Experiments extends AbstractEntity
      */
     public function duplicate(): int
     {
-        // anon cannot duplicate anything
-        if (isset($this->Users->userData['anon']) && $this->Users->userData['anon'] === true) {
-            throw new Exception(Tools::error(true));
-        }
-
         // let's add something at the end of the title to show it's a duplicate
         // capital i looks good enough
         $title = $this->entityData['title'] . ' I';
@@ -247,7 +258,7 @@ class Experiments extends AbstractEntity
             'status' => $this->getStatus(),
             'elabid' => $this->generateElabid(),
             'visibility' => $this->entityData['visibility'],
-            'userid' => $this->Users->userid));
+            'userid' => $this->Users->userData['userid']));
         $newId = $this->Db->lastInsertId();
 
         $this->Links->duplicate($this->id, $newId);
@@ -260,42 +271,70 @@ class Experiments extends AbstractEntity
     /**
      * Destroy an experiment and all associated data
      *
-     * @return bool
+     * @return void
      */
-    public function destroy(): bool
+    public function destroy(): void
     {
+        $this->canOrExplode('write');
+
         $sql = "DELETE FROM experiments WHERE id = :id";
         $req = $this->Db->prepare($sql);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
-        $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
 
         $this->Comments->destroyAll();
         $this->Links->destroyAll();
         $this->Steps->destroyAll();
         $this->Tags->destroyAll();
         $this->Uploads->destroyAll();
+    }
 
-        return true;
+    public function getTimestampInfo(): array
+    {
+        if ($this->entityData['timestamped'] === '0') {
+            return array();
+        }
+        $timestamper = $this->Users->read((int) $this->entityData['timestampedby']);
+
+        $Uploads = new Uploads(new Experiments($this->Users, (int) $this->entityData['id']));
+        $Uploads->Entity->type = 'exp-pdf-timestamp';
+        $pdf = $Uploads->readAll();
+
+        $Uploads->Entity->type = 'timestamp-token';
+        $token = $Uploads->readAll();
+
+        return array(
+            'timestamper' => $timestamper,
+            'pdf' => $pdf,
+            'token' => $token
+        );
     }
 
     /**
      * Lock/unlock
      *
-     * @throws Exception
-     * @return bool
+     * @return void
      */
-    public function toggleLock(): bool
+    public function toggleLock(): void
     {
+        $permissions = $this->getPermissions();
+        if (!$this->Users->userData['can_lock'] && !$permissions['write']) {
+            throw new ImproperActionException(_("You don't have the rights to lock/unlock this."));
+        }
         $locked = (int) $this->entityData['locked'];
 
         // if we try to unlock something we didn't lock
-        if ($locked === 1 && ($this->entityData['lockedby'] != $this->Users->userid)) {
+        if ($locked === 1 && ($this->entityData['lockedby'] != $this->Users->userData['userid'])) {
             // Get the first name of the locker to show in error message
             $sql = "SELECT firstname FROM users WHERE userid = :userid";
             $req = $this->Db->prepare($sql);
             $req->bindParam(':userid', $this->entityData['lockedby'], PDO::PARAM_INT);
-            $req->execute();
-            throw new Exception(
+            if ($req->execute() !== true) {
+                throw new DatabaseErrorException('Error while executing SQL query.');
+            }
+            throw new ImproperActionException(
                 _('This experiment was locked by') .
                 ' ' . $req->fetchColumn() . '. ' .
                 _("You don't have the rights to unlock this.")
@@ -304,22 +343,16 @@ class Experiments extends AbstractEntity
 
         // check if the experiment is timestamped. Disallow unlock in this case.
         if ($locked === 1 && $this->entityData['timestamped']) {
-            throw new Exception(_('You cannot unlock or edit in any way a timestamped experiment.'));
+            throw new ImproperActionException(_('You cannot unlock or edit in any way a timestamped experiment.'));
         }
 
-        // toggle
-        if ($locked === 1) {
-            $locked = 0;
-        } else {
-            $locked = 1;
-        }
-        $sql = "UPDATE experiments
-            SET locked = :locked, lockedby = :lockedby, lockedwhen = CURRENT_TIMESTAMP WHERE id = :id";
+        $sql = "UPDATE experiments SET locked = IF(locked = 1, 0, 1), lockedby = :lockedby, lockedwhen = CURRENT_TIMESTAMP WHERE id = :id";
         $req = $this->Db->prepare($sql);
-        $req->bindParam(':locked', $locked, PDO::PARAM_INT);
-        $req->bindParam(':lockedby', $this->Users->userid, PDO::PARAM_INT);
+        $req->bindParam(':lockedby', $this->Users->userData['userid'], PDO::PARAM_INT);
         $req->bindParam(':id', $this->id, PDO::PARAM_INT);
 
-        return $req->execute();
+        if ($req->execute() !== true) {
+            throw new DatabaseErrorException('Error while executing SQL query.');
+        }
     }
 }
